opacityUsingHoverAndRgb.html_12_08_22
The opacity property is often used together with the :hover selector to change the opacity on mouse-over:When using the opacity property to add transparency to the background of an element, all of its child elements inherit the same transparency. This can make the text inside a fully transparent element hard to read:If you do not want to apply opacity to child elements, like in our example above, use RGBA color values.