flex-direction_17_08_22.html
Even though you haven't added a flex-direction property yet, the items display as a row because the initial value of flex-direction is row. If you want a row then you don't need to add the property. To change the direction, add the property and one of the four values.row: the items lay out as a row.  row-reverse: the items lay out as a row from the end of the flex container. column: the items lay out as a column. column-reverse : the items lay out as a column from the end of the flex container.